[#8] Refactor UUID creation to BaseModel class
